完善pdd自动任务订单保存

+++ serviceImpl/order/src/main/java/cn/qihangerp/module/order/service/OOrderService.java
@@ -25,7 +25,7 @@ public interface OOrderService extends IService<OOrder> {
     ResultVo<Integer> jdOrderMessage(String orderId, JSONObject orderDetail);
 
     ResultVo<Integer> taoOrderMessage(String tid,JSONObject orderDetail );
-    ResultVo<Integer> pddOrderMessage(String orderSn,JSONObject orderDetail );
+    ResultVo<Long> pddOrderMessage(String orderSn,JSONObject orderDetail );
     ResultVo<Integer> douOrderMessage(String orderId,JSONObject orderDetail );
     ResultVo<Integer> offlineOrderMessage(String orderNum);
     ResultVo<Integer> weiOrderMessage(String orderNum,JSONObject orderDetail );
